changed ActivationFunction* to ActivationFunction in models
Former-commit-id: dd5be6cb5873855375f9cc2a59dd8f4d5c596ebb [formerly f0bc45f778c8622247d2b284e0b5c1c025bb3f5a]
Former-commit-id: 4ca8d553a8b3958c81d3a966851f3c3f54f48ef9

+++ src/gui/utils/ModelTrainWorker.cpp
@@ -18,7 +18,7 @@ void ModelTrainWorker::process() {
 					return;
 				if (j != (size_t) dataset.current_iteration || dataset.size() == 1) {
 					correct_guesses = 0;
-					for (int k = 0; k < (int) dataset[j].size(); ++k) {
+					for (size_t k = 0; k < dataset[j].size(); ++k) {
 						if (this->stop)
 							return;
